Eliminado de los formularios lo referido a intervenciones y creado su propio form en la app intervenciones
